Update modifiers to be more option-oriented
The 'groups' are now high-level specifiers of the approach the modifier
takes to using its options.

The options themselves now influence the matching structure much more,
as the expressions (and modifiers) can check the presence of options and
change their behavior accordingly.

Still a few things to clear up here - especially with the way options
are passed down through expressions.